Correccion en test y se elimino clase CalculadorReparacion porque estaba duplicada
