+++ src/test/java/com/grupo4/trabajo/Servicios/RechazarPedidoEconomicPorDeuda.java
@@ -3,6 +3,7 @@ package com.grupo4.trabajo.Servicios;
 import com.grupo4.trabajo.Cliente;
 import com.grupo4.trabajo.Exceptions.EsDeudorException;
 import com.grupo4.trabajo.Pedido;
+import com.grupo4.trabajo.Servicios.ServicioCliente.Economic;
 import org.junit.jupiter.api.Test;
 import com.grupo4.trabajo.Robots.*;
